Create Stand_03.stl
Adding a stand that uses a dead CR2023 battery to display a Blink on a desk, much like a little trophy. Note, the bottom of the stand can be linked together with many others and works to display a whole set of blinks in a very neat way. Inspired by one shared in the community forum.